Address PR comments and fix blog back button
