+++ frontend/e2e/smoke/vendor-directory.public.spec.ts
@@ -23,7 +23,7 @@ test.describe('Vendor directory — guest', () => {
   });
 
   test.beforeEach(async ({ page }) => {
-    await page.goto('/directory');
+    await page.goto('/vendors');
     // Wait until the result-count line renders, confirming vendors are loaded
     await expect(page.getByText(/4 Wedding Beauty Artists found/)).toBeVisible({ timeout: 15_000 });
   });
